ADM-fix-37a9.17: Updated maintenance script

+++ admin/ajax_maintenance.php
@@ -29,7 +29,7 @@ $ques = array(
   'DELETE FROM `{{annonce}}`   WHERE `user`           not in (select id from {{users}});',
 //  'DELETE FROM `{{messages}}`  WHERE `message_sender` not in (select id from {{users}});',
   'DELETE FROM `{{messages}}`  WHERE `message_owner`  not in (select id from {{users}});',
-  'DELETE FROM `{{planets}}`   WHERE `id_owner`       not in (select id from {{users}});',
+  'DELETE FROM `{{planets}}`   WHERE `id_owner`       not in (select id from {{users}}) AND id_owner <> 0;',
 
   // UBE reports
   'DELETE FROM `{{ube_report}}` WHERE `ube_report_time_combat` < DATE_SUB(now(),INTERVAL 60 day);',
